Editted Current Files
Editted the PlayerHandler script to have multiple menus and options. Skills was just naming changes. Spells in a new script which will end up similar to skills. PlayerStats now has a particle system which will play when the player levels up and the script was changed to accomodate spell points